+++ server/weapons/rpg.cpp
@@ -22,6 +22,7 @@
 #include "player.h"
 #include "gamerules.h"
 #include "user_messages.h"
+#include "ggrenade.h"
 
 enum rpg_e
 {
